+++ spring-cloud-starter-tencent-polaris-circuitbreaker/src/main/java/com/tencent/cloud/polaris/circuitbreaker/util/PolarisCircuitBreakerUtils.java
@@ -1,0 +1,54 @@
+/*
+ * Tencent is pleased to support the open source community by making Spring Cloud Tencent available.
+ *
+ * Copyright (C) 2019 THL A29 Limited, a Tencent company. All rights reserved.
+ *
+ * Licensed under the BSD 3-Clause License (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * https://opensource.org/licenses/BSD-3-Clause
+ *
+ * Unless required by applicable law or agreed to in writing, software distributed
+ * under the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR
+ * CONDITIONS OF ANY KIND, either express or implied. See the License for the
+ * specific language governing permissions and limitations under the License.
+ */
+
+package com.tencent.cloud.polaris.circuitbreaker.util;
+
+import com.tencent.cloud.common.metadata.MetadataContext;
+
+import org.springframework.util.Assert;
+
+/**
+ * PolarisCircuitBreakerUtils.
+ *
+ * @author seanyu 2023-02-27
+ */
+public final class PolarisCircuitBreakerUtils {
+
+	private PolarisCircuitBreakerUtils() {
+
+	}
+
+	/**
+	 *
+	 * @param id CircuitBreakerId
+	 *              Format: namespace#service#method or service#method or service ,
+	 *              namespace set as default spring.cloud.polaris.namespace if absent
+	 * @return String[]{namespace, service, method}
+	 */
+	public static String[] resolveCircuitBreakerId(String id) {
+		Assert.hasText(id, "A CircuitBreaker must have an id. Id could be : namespace#service#method or service#method or service");
+		String[] polarisCircuitBreakerMetaData = id.split("#");
+		if (polarisCircuitBreakerMetaData.length == 2) {
+			return new String[]{MetadataContext.LOCAL_NAMESPACE, polarisCircuitBreakerMetaData[0], polarisCircuitBreakerMetaData[1]};
+		}
+		if (polarisCircuitBreakerMetaData.length == 3) {
+			return new String[]{polarisCircuitBreakerMetaData[0], polarisCircuitBreakerMetaData[1], polarisCircuitBreakerMetaData[2]};
+		}
+		return new String[]{MetadataContext.LOCAL_NAMESPACE, id, ""};
+	}
+
+}