+++ CHANGELOG.md
@@ -12,3 +12,4 @@
 - [feature:add User-Agent:polaris for healthyCheck api.](https://github.com/Tencent/spring-cloud-tencent/pull/873)
 - [optimize:optimize ServiceRuleManager.](https://github.com/Tencent/spring-cloud-tencent/pull/878)
 - [refactor:refactor stat module.](https://github.com/Tencent/spring-cloud-tencent/pull/879)
+- [feature:add polaris circuit breaker support.](https://github.com/Tencent/spring-cloud-tencent/pull/881)